Remove abstract class TimeSeriesChartD3
