+++ src/d3/line-chart.d3.ts
@@ -1,7 +1,8 @@
-import { TimeSeriesChartD3 } from './time-series-chart.d3';
 import * as d3 from 'd3';
+import { XYChartD3 } from './xy-chart.d3';
 import { LineChartDatum } from '../components/line-chart/line-chart.component';
 import { TimeSeriesPoint } from '../datasets/metas/types';
+import { formatX } from '../utils/formatters';
 
 export interface LegendItemStyle {
   color: string;
@@ -15,7 +16,7 @@ interface LinePath {
   path: d3.Selection<SVGPathElement, unknown, null, undefined>;
 }
 
-export class LineChartD3 extends TimeSeriesChartD3<LegendItemStyle> {
+export class LineChartD3 extends XYChartD3<TimeSeriesPoint, LineChartDatum> {
   static defaultLegendItemStyle: LegendItemStyle = {
     color: LineChartD3.primaryColor,
     width: 2,
@@ -25,6 +26,19 @@ export class LineChartD3 extends TimeSeriesChartD3<LegendItemStyle> {
 
   protected linePaths: LinePath[];
   protected activePointCircle: d3.Selection<SVGCircleElement, unknown, null, undefined>;
+
+  protected scaleX: d3.ScaleTime<number, number>;
+  protected scaleY: d3.ScaleLinear<number, number>;
+  protected xAxis: d3.Axis<Date>;
+  protected yAxis: d3.Axis<number>;
+
+  x(value: Date) {
+    return this.scaleX(value);
+  }
+
+  y(value: number) {
+    return this.scaleY(value);
+  }
 
   protected renderData() {
     super.renderData();
@@ -68,6 +82,56 @@ export class LineChartD3 extends TimeSeriesChartD3<LegendItemStyle> {
     });
   }
 
+  protected renderAxis() {
+    const { height, width } = this.renderOptions;
+    const { xAxisHeight, yAxisWidth, legendHeight, padding, fontSizeSmall } = XYChartD3;
+
+    const top = padding;
+    const bottom = height - xAxisHeight - legendHeight - padding;
+    const left = padding + yAxisWidth;
+    const right = width - padding;
+
+    this.scaleX = d3
+      .scaleUtc()
+      .range([left, right]);
+    this.scaleY = d3
+      .scaleLinear()
+      .nice()
+      .range([bottom, top]);
+
+    this.xAxis = d3
+      .axisBottom<Date>(this.scaleX)
+      .ticks(d3.timeWeek.every(1))
+      .tickFormat(formatX);
+    this.yAxis = d3.axisLeft<number>(this.scaleY);
+
+    this.xAxisG = this.svg
+      .append('g')
+      .attr('class', 'xy_chart-x_axis')
+      .attr('transform', `translate(0,${bottom})`)
+      .attr('font-size', fontSizeSmall);
+    this.yAxisG = this.svg
+      .append('g')
+      .attr('class', 'xy_chart-y_axis')
+      .attr('transform', `translate(${left},0)`)
+      .attr('font-size', fontSizeSmall);
+  }
+
+  protected updateAxis(data: LineChartDatum[]) {
+    const points = data.flatMap(datum => datum.points);
+    const xs = points.map(point => point.x);
+    const ys = points.map(point => point.y);
+    this.scaleX.domain([d3.min(xs)!, d3.max(xs)!]);
+    this.scaleY.domain([0, d3.max(ys)!]);
+
+    this.xAxisG
+      .transition(this.transition)
+      .call(this.xAxis);
+    this.yAxisG
+      .transition(this.transition)
+      .call(this.yAxis);
+  }
+
   protected renderActivePoint() {
     this.activePointCircle = this.svg
       .append('circle')
@@ -90,7 +154,7 @@ export class LineChartD3 extends TimeSeriesChartD3<LegendItemStyle> {
   }
 
   protected appendLegendItemIcon(container: d3.Selection<SVGGElement, unknown, null, undefined>, datum: LineChartDatum) {
-    const { legendIconWidth, legendHeight } = TimeSeriesChartD3;
+    const { legendIconWidth, legendHeight } = XYChartD3;
     container
       .append('line')
       .attr('x1', 0)
